+++ packages/@biomejs/js-api/src/web.ts
@@ -1,5 +1,5 @@
-import type { Configuration, Diagnostic } from "@blazing-fast-rust/wasm-web";
-import * as moduleWeb from "@blazing-fast-rust/wasm-web";
+import type { Configuration, Diagnostic } from "blazing-fast-rust-wasm-web";
+import * as moduleWeb from "blazing-fast-rust-wasm-web";
 import { blazing-fast-rustCommon } from "./common";
 
 export type * from "./common";