+++ packages/@biomejs/cli-darwin-x64/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @blazing-fast-rust/cli-darwin-x64
+# blazing-fast-rust-cli-darwin-x64
 
 ## 2.1.2